+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * **(breaking change)** `find()` now distinguishes between files and directories and by default searches only for files (previously searched for both).
 * **(breaking change)** `list()` no longer accepts `useInspect` as a parameter. To achieve old behaviour use `jetpack.list()` with `Array.map()`.
 * **(deprecation)** Don't do `jetpack.read('sth', 'buf')`, do `jetpack.read('sth', 'buffer')` instead.
-* `remove()`, `list()` can be now called without any parameters and defaults to CWD in that case.
+* `remove()`, `list()` and `find()` can be now called without provided `path`, and defaults to CWD in that case.
 
 0.7.3 (2016-03-21)
 -------------------