+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 -------------------
 * **(breaking change)** `find()` now distinguishes between files and directories and by default searches only for files (previously searched for both).
 * **(breaking change)** `list()` no longer accepts `useInspect` as a parameter. To achieve old behaviour use `jetpack.list()` with `Array.map()`.
+* **(breaking change)** `find()` no longer accepts `'inspect'` as `returnAs` parameter.
 * **(deprecation)** Don't do `jetpack.read('sth', 'buf')`, do `jetpack.read('sth', 'buffer')` instead.
 * `remove()`, `list()` and `find()` can be now called without provided `path`, and defaults to CWD in that case.